Fix build: disambiguate volume drag math (compute fraction in Double)
VolumeControl's drag snapped via (frac*100).rounded()/100 where frac was CGFloat,
making the operator/Double conversion ambiguous for onChange(Double). Compute the
fraction as Double up front. (Sole compile error in f5bdb0a; badges already sit
on their own row above the transport row, so the volume expansion can't obscure
them in inline or fullscreen.)

+++ ios/Stashy/Features/Player/VolumeControl.swift
@@ -70,7 +70,7 @@ struct VolumeControl: View {
                 DragGesture(minimumDistance: 0)
                     .onChanged { v in
                         // Snap to whole-percent steps so the value moves 0–100 in 1-unit increments.
-                        let frac = min(1, max(0, v.location.x / w))
+                        let frac = Double(min(1, max(0, v.location.x / w)))
                         onChange((frac * 100).rounded() / 100)
                         onInteract()
                         armCollapse()
